+++ ICD.Connect.Analytics.FusionPro/Properties/AssemblyInfo.cs
@@ -1,7 +1,0 @@
-using System.Reflection;
-
-[assembly: AssemblyTitle("ICD.Connect.Analytics.FusionPro")]
-[assembly: AssemblyCompany("ICD Systems")]
-[assembly: AssemblyProduct("ICD.Connect.Analytics.FusionPro")]
-[assembly: AssemblyCopyright("Copyright © ICD Systems 2017")]
-[assembly: AssemblyVersion("3.0.1.0")]