+++ ICD.Connect.Analytics/Properties/AssemblyInfo.cs
@@ -1,8 +1,0 @@
-using System.Reflection;
-
-[assembly: AssemblyTitle("ICD.Connect.Analytics")]
-[assembly: AssemblyCompany("ICD Systems")]
-[assembly: AssemblyProduct("ICD.Connect.Analytics")]
-[assembly: AssemblyCopyright("Copyright © ICD Systems 2017")]
-[assembly: AssemblyVersion("3.0.1.0")]
-