+++ ICD.Connect.Telemetry.CrestronPro/Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+using System.Reflection;
+
+[assembly: AssemblyTitle("ICD.Connect.Telemetry.CrestronPro")]
+[assembly: AssemblyCompany("ICD Systems")]
+[assembly: AssemblyProduct("ICD.Connect.Telemetry.CrestronPro")]
+[assembly: AssemblyCopyright("Copyright © ICD Systems 2018")]
+[assembly: AssemblyVersion("4.0.0.0")]