+++ ICD.Connect.Telemetry/Properties/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+using System.Reflection;
+
+[assembly: AssemblyTitle("ICD.Connect.Telemetry")]
+[assembly: AssemblyCompany("ICD Systems")]
+[assembly: AssemblyProduct("ICD.Connect.Telemetry")]
+[assembly: AssemblyCopyright("Copyright © ICD Systems 2018")]
+[assembly: AssemblyVersion("4.0.0.0")]
+