+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [4.0.0] - 2018-10-18
+### Changed
+ - Analytics renamed to Telemetry
+ - FusionPro renamed to CrestronPro
+
+## [3.0.3] - 2018-09-14
+### Changed
+ - Fixing using directives
+
+## [3.0.2] - 2018-07-02
+### Changed
+ - Sig collections are now public
+
 ## [3.0.1] - 2018-05-09
 ### Changed
  - Fixing issues with loading UTF8-BOM configs